reenable rc e2e
