+++ test/e2e/manifests/new_manifests/datadog-agent-rc-updater.yaml
